+++ src/image.py
@@ -6,13 +6,16 @@
 import json
 import logging
 from collections import defaultdict
-from typing import Mapping, TypedDict, cast
+from typing import TYPE_CHECKING, Mapping, TypedDict, cast
 
 import ops
 
 import builder
 import charm_utils
 import state
+
+if TYPE_CHECKING:
+    from charm import GithubRunnerImageBuilderCharm
 
 logger = logging.getLogger(__name__)
 
@@ -34,7 +37,7 @@ class ImageRelationData(TypedDict, total=False):
 class Observer(ops.Object):
     """The image relation observer."""
 
-    def __init__(self, charm: ops.CharmBase):
+    def __init__(self, charm: "GithubRunnerImageBuilderCharm"):
         """Initialize the observer and register event handlers.
 
         Args:
@@ -55,6 +58,7 @@ class Observer(ops.Object):
             event: The event emitted when a relation is joined.
         """
         build_config = state.BuilderConfig.from_charm(charm=self.charm)
+        self.charm.setup_proxy_environment(build_config.proxy)
         proxy = state.ProxyConfig.from_env()
         if not build_config.cloud_config.upload_cloud_ids:
             self.model.unit.status = ops.BlockedStatus(