+++ docs/changelog.md
@@ -1,6 +1,9 @@
 
 <!-- vale Canonical.007-Headings-sentence-case = NO -->
 
+## [#213 Fix proxy setup]
+* Fix proxy setup for image-relation joined hook.
+
 ## [#219 Use Juju secrets](https://github.com/canonical/github-runner-image-builder-operator/pull/219) (2026-04-17)
 * Add new `openstack-password-secret` configuration option to securely store OpenStack passwords using Juju secrets.
 * Deprecated `openstack-password` configuration option (still supported for backward compatibility).